money: store and compute money as exact decimal, not binary float (D-58) (#82)
The probe that opened this row, now a test: creating a $1,000 income on an
account holding 1104.55 and deleting it again left 1104.5500000000002. The size
is not the point — Account.balance is mutated in place and nothing ever
re-derives it from the transactions that produced it, so the error accumulates
and never self-corrects, and that balance is what the Accounts page sums into net
worth.

The row's own proposed shortcut was checked and is impossible: "recompute from
the rows" needs an opening balance, and Account has no such column. So the
columns had to change.

20 columns move to Numeric, in three scales chosen per cluster rather than
uniformly: 18,2 for currency, 20,8 for share counts, 20,10 for exchange rates.
import_sources.confidence and import_batches.confidence stay Float deliberately —
0..1 heuristic scores that never touch money — and say so in the model.

Partial conversion inside one arithmetic cluster is worse than none, because
Decimal * float raises rather than converting. Proven while probing: reverting
Account.balance alone turned three passing tests into 500s. So the boundaries are
explicit and singular. src/utils/money.py is the one place an outside number
becomes money, and it coerces through str() on purpose — Decimal(0.1) preserves
the binary error faithfully, Decimal(str(0.1)) discards it.

Fixed at every site the suite reached, each read first: the balance move, the
split arithmetic (percentage and custom branches, plus the 0.01 tolerances), the
budget accumulator that started at 0.0, the analytics monthly and investment
accumulators, and the four rule-of-thumb ratios in financial health that
multiplied a total by a float literal.

Two encoders had to learn Decimal, and only one is obvious. Flask's JSON provider
covers jsonify; flask-restx does NOT use it — output_json calls the stdlib dumps
directly with whatever is in RESTX_JSON — so setting only the first leaves every
/api/v1 response 500ing. Both are wired, both emit a JSON number rather than a
string, and that is pinned on the raw body text because it is a client-visible
contract across two apps, not an implementation detail.

The gate is keyed to `not isinstance(type, Float)`, NOT to `isinstance(...,
Numeric)`: in SQLAlchemy Float is a subclass of Numeric, so the obvious spelling
would have certified success on a column that had not been converted at all —
the vacuous-check failure mode, avoided by writing it the other way and probing
it.

The migration is also a merge point. The revision tree had three heads, so
`alembic upgrade head` could not run without naming one; depending on all three
fixes that rather than adding a fourth. It carries the warning that create_all()
does not alter existing columns, so on an existing database the model change
alone is a silent no-op.

backend 958 -> 964.

+++ src/models/group.py
@@ -155,7 +155,7 @@ class Settlement(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     payer_id = db.Column(db.String(120), db.ForeignKey('users.id'), nullable=False)
     receiver_id = db.Column(db.String(120), db.ForeignKey('users.id'), nullable=False)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(18, 2), nullable=False)
     date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
     description = db.Column(db.String(200), nullable=True, default="Settlement")
